requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
boto3==1.29.7
PyPDF2==3.0.1
python-docx==1.1.0
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0